move override logic to new lib folder
