+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "console-to-slack",
-  "version": "1.1.0-pre-josh-0",
+  "version": "1.1.0-pre-josh-1",
   "description": "Module for integrating console usage with Slack webhooks.",
   "main": "index.js",
   "engines": {
